CurrencyExchange/requirements.txt
redis
python-dotenv
aiohttp